Removed CDW support for MedicationStatementController
