+++ data-query-tests/src/test/resources/cdw/datamart-index.yaml
@@ -284,46 +284,6 @@ entries:
     query: /MedicationDispense:1.02?patient=43000199&type=FF,UD
     page: 1
     count: 15
-  - file: ../../../../target/cdw/samples/MedSta102i1000000398707.xml
-    query: /MedicationStatement:1.02?identifier=1000000398707
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedSta102i1400000182116.xml
-    query: /MedicationStatement:1.02?identifier=800000000707
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedSta102i5555555555555.xml
-    query: /MedicationStatement:1.02?identifier=5555555555555
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 1
-    count: 15
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg0.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 0
-    count: 15
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg10000ct20.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 10000
-    count: 20
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg1ct0.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 1
-    count: 0
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg1ct2.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 1
-    count: 2
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg1ct21.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 1
-    count: 21
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290pg2ct2.xml
-    query: /MedicationStatement:1.02?patient=43000199
-    page: 2
-    count: 2
   - file: ../../../../target/cdw/samples/Obs102catlaboratoryd20150415p185601V825290.xml
     query: /Observation:1.02?category=laboratory&date=2015-04-15&patient=43000199
     page: 1
@@ -770,19 +730,6 @@ entries:
     count: 1
   - file: ../../../../target/cdw/samples/Enc101p185601V825290.xml
     query: /Encounter:1.02?_id=1200753214085
-    page: 1
-    count: 1
-    # MedicationStatement aliases
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290.xml
-    query: /MedicationStatement:1.02?identifier=800000000707
-    page: 1
-    count: 1
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290.xml
-    query: /MedicationStatement:1.02?_id=800000000707
-    page: 1
-    count: 1
-  - file: ../../../../target/cdw/samples/MedSta102p185601V825290.xml
-    query: /MedicationStatement:1.02?patient=43000199&_include=MedicationStatement:medication
     page: 1
     count: 1
     # DiagnosticReport aliases
