+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/medicationstatement/Dstu2MedicationStatementControllerTest.java
@@ -14,13 +14,14 @@ import gov.va.api.health.dataquery.service.controller.Bundler;
 import gov.va.api.health.dataquery.service.controller.ConfigurableBaseUrlPageLinks;
 import gov.va.api.health.dataquery.service.controller.ResourceExceptions;
 import gov.va.api.health.dataquery.service.controller.WitnessProtection;
-import gov.va.api.health.dataquery.service.controller.medicationstatement.DatamartMedicationStatementSamples.Datamart;
-import gov.va.api.health.dataquery.service.controller.medicationstatement.DatamartMedicationStatementSamples.Fhir;
+import gov.va.api.health.dataquery.service.controller.medicationstatement.MedicationStatementSamples.Datamart;
+import gov.va.api.health.dataquery.service.controller.medicationstatement.MedicationStatementSamples.Dstu2;
 import gov.va.api.health.dstu2.api.bundle.BundleLink.LinkRelation;
 import gov.va.api.health.ids.api.IdentityService;
 import gov.va.api.health.ids.api.Registration;
 import gov.va.api.health.ids.api.ResourceIdentity;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import javax.servlet.http.HttpServletResponse;
 import lombok.SneakyThrows;
@@ -35,7 +36,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @DataJpaTest
 @RunWith(SpringRunner.class)
-public class DatamartMedicationStatementControllerTest {
+public class Dstu2MedicationStatementControllerTest {
 
   HttpServletResponse response;
 
@@ -59,11 +60,8 @@ public class DatamartMedicationStatementControllerTest {
         .build();
   }
 
-  MedicationStatementController controller() {
-    return new MedicationStatementController(
-        true,
-        null,
-        null,
+  Dstu2MedicationStatementController controller() {
+    return new Dstu2MedicationStatementController(
         new Bundler(new ConfigurableBaseUrlPageLinks("http://fonzy.com", "cool")),
         repository,
         WitnessProtection.builder().identityService(ids).build());
@@ -89,7 +87,7 @@ public class DatamartMedicationStatementControllerTest {
   }
 
   private Multimap<String, MedicationStatement> populateData() {
-    var fhir = Fhir.create();
+    var fhir = Dstu2.create();
     var datamart = Datamart.create();
     var medicationStatementByPatient = LinkedHashMultimap.<String, MedicationStatement>create();
     var registrations = new ArrayList<Registration>(10);
@@ -119,17 +117,17 @@ public class DatamartMedicationStatementControllerTest {
   @Test
   public void read() {
     DatamartMedicationStatement dm =
-        DatamartMedicationStatementSamples.Datamart.create().medicationStatement();
+        MedicationStatementSamples.Datamart.create().medicationStatement();
     repository.save(asEntity(dm));
     mockMedicationStatementIdentity("1", dm.cdwId());
-    MedicationStatement actual = controller().read("true", "1");
-    assertThat(json(actual)).isEqualTo(json(Fhir.create().medicationStatement("1")));
+    MedicationStatement actual = controller().read("1");
+    assertThat(json(actual)).isEqualTo(json(Dstu2.create().medicationStatement("1")));
   }
 
   @Test
   public void readRaw() {
     DatamartMedicationStatement dm =
-        DatamartMedicationStatementSamples.Datamart.create().medicationStatement();
+        MedicationStatementSamples.Datamart.create().medicationStatement();
     MedicationStatementEntity entity = asEntity(dm);
     repository.save(entity);
     mockMedicationStatementIdentity("1", dm.cdwId());
@@ -152,12 +150,12 @@ public class DatamartMedicationStatementControllerTest {
   @Test(expected = ResourceExceptions.NotFound.class)
   public void readThrowsNotFoundWhenDataIsMissing() {
     mockMedicationStatementIdentity("1", "1");
-    controller().read("true", "1");
+    controller().read("1");
   }
 
   @Test(expected = ResourceExceptions.NotFound.class)
   public void readThrowsNotFoundWhenIdIsUnknown() {
-    controller().read("true", "1");
+    controller().read("1");
   }
 
   @Test
@@ -165,30 +163,17 @@ public class DatamartMedicationStatementControllerTest {
     DatamartMedicationStatement dm = Datamart.create().medicationStatement();
     repository.save(asEntity(dm));
     mockMedicationStatementIdentity("1", dm.cdwId());
-    Bundle actual = controller().searchById("true", "1", 1, 1);
-    MedicationStatement medicationStatement =
-        Fhir.create().medicationStatement("1", dm.patient().reference().get());
-    assertThat(json(actual))
-        .isEqualTo(
-            json(
-                Fhir.asBundle(
-                    "http://fonzy.com/cool",
-                    List.of(medicationStatement),
-                    Fhir.link(
-                        LinkRelation.first,
-                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
-                        1,
-                        1),
-                    Fhir.link(
-                        LinkRelation.self,
-                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
-                        1,
-                        1),
-                    Fhir.link(
-                        LinkRelation.last,
-                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
-                        1,
-                        1))));
+    Bundle actual = controller().searchById("1", 1, 1);
+    validateSearchByIdResult(dm, actual);
+  }
+
+  @Test
+  public void searchByIdentifier() {
+    DatamartMedicationStatement dm = Datamart.create().medicationStatement();
+    repository.save(asEntity(dm));
+    mockMedicationStatementIdentity("1", dm.cdwId());
+    Bundle actual = controller().searchByIdentifier("1", 1, 1);
+    validateSearchByIdResult(dm, actual);
   }
 
   @Test
@@ -197,28 +182,73 @@ public class DatamartMedicationStatementControllerTest {
     assertThat(json(controller().searchByPatient("true", "p0", 1, 10)))
         .isEqualTo(
             json(
-                Fhir.asBundle(
+                Dstu2.asBundle(
                     "http://fonzy.com/cool",
                     medicationStatementByPatient.get("p0"),
-                    Fhir.link(
+                    medicationStatementByPatient.get("p0").size(),
+                    Dstu2.link(
                         LinkRelation.first,
                         "http://fonzy.com/cool/MedicationStatement?patient=p0",
                         1,
                         10),
-                    Fhir.link(
+                    Dstu2.link(
                         LinkRelation.self,
                         "http://fonzy.com/cool/MedicationStatement?patient=p0",
                         1,
                         10),
-                    Fhir.link(
+                    Dstu2.link(
                         LinkRelation.last,
                         "http://fonzy.com/cool/MedicationStatement?patient=p0",
                         1,
                         10))));
   }
 
+  @Test
+  public void searchByPatientWithCount0() {
+    Multimap<String, MedicationStatement> medicationStatementByPatient = populateData();
+    assertThat(json(controller().searchByPatient("true", "p0", 1, 0)))
+        .isEqualTo(
+            json(
+                Dstu2.asBundle(
+                    "http://fonzy.com/cool",
+                    Collections.emptyList(),
+                    medicationStatementByPatient.get("p0").size(),
+                    Dstu2.link(
+                        LinkRelation.self,
+                        "http://fonzy.com/cool/MedicationStatement?patient=p0",
+                        1,
+                        0))));
+  }
+
   @SneakyThrows
   private DatamartMedicationStatement toObject(String json) {
     return JacksonConfig.createMapper().readValue(json, DatamartMedicationStatement.class);
+  }
+
+  private void validateSearchByIdResult(DatamartMedicationStatement dm, Bundle actual) {
+    MedicationStatement medicationStatement =
+        Dstu2.create().medicationStatement("1", dm.patient().reference().get());
+    assertThat(json(actual))
+        .isEqualTo(
+            json(
+                Dstu2.asBundle(
+                    "http://fonzy.com/cool",
+                    List.of(medicationStatement),
+                    1,
+                    Dstu2.link(
+                        LinkRelation.first,
+                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
+                        1,
+                        1),
+                    Dstu2.link(
+                        LinkRelation.self,
+                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
+                        1,
+                        1),
+                    Dstu2.link(
+                        LinkRelation.last,
+                        "http://fonzy.com/cool/MedicationStatement?identifier=1",
+                        1,
+                        1))));
   }
 }
